renamed fixtures into conftest and created pytest.ini
