+++ pytest.ini
@@ -1,0 +1,6 @@
+[pytest]
+testpaths = /tests
+norecursedirs = .* build dist venv env .venv __pycache__ site-packages
+filterwarnings =
+    ignore::DeprecationWarning:werkzeug.*
+    ignore::DeprecationWarning:ast.*
